@@ -1,9 +1,9 @@
-import Card from "../components/card";
-import CustomHead from "../components/CustomHead";
-import { useState, useEffect } from "react";
-const allTheWordsAvailable = require("../words.json");
+import Card from '../components/card';
+import CustomHead from '../components/CustomHead';
+import { useState, useEffect } from 'react';
+const allTheWordsAvailable = require('../words.json');
 
-const generateWords = (words) => {
+const generateWords = words => {
   const pickedWords = [];
   const wordsCloned = Object.assign([], words);
   for (let i = 0; i < 16; i++) {
@@ -31,26 +31,26 @@ const Index = () => {
   const [isTeam1, setIsTeam1] = useState(true);
 
   useEffect(() => {
-    setTeam1Points(getState("team1Points", 0));
-    setTeam2Points(getState("team2Points", 0));
+    setTeam1Points(getState('team1Points', 0));
+    setTeam2Points(getState('team2Points', 0));
     setCurrentWords(
-      getState("currentWords", generateWords(allTheWordsAvailable))
+      getState('currentWords', generateWords(allTheWordsAvailable))
     );
-    setIsTeam1(getState("isTeam1", true));
+    setIsTeam1(getState('isTeam1', true));
     setIsLoading(false);
   }, []);
 
   useEffect(() => {
-    localStorage.setItem("team1Points", JSON.stringify(team1Points));
+    localStorage.setItem('team1Points', JSON.stringify(team1Points));
   }, [team1Points]);
   useEffect(() => {
-    localStorage.setItem("team2Points", JSON.stringify(team2Points));
+    localStorage.setItem('team2Points', JSON.stringify(team2Points));
   }, [team2Points]);
   useEffect(() => {
-    localStorage.setItem("currentWords", JSON.stringify(currentWords));
+    localStorage.setItem('currentWords', JSON.stringify(currentWords));
   }, [currentWords, currentWords.length]);
   useEffect(() => {
-    localStorage.setItem("isTeam1", JSON.stringify(isTeam1));
+    localStorage.setItem('isTeam1', JSON.stringify(isTeam1));
   }, [isTeam1]);
 
   const restart = () => {
@@ -80,7 +80,7 @@ const Index = () => {
             <div className="scoring">
               <div
                 className={`score
-          ${currentWords.length > 0 && isTeam1 ? "team-active" : ""}`}
+          ${currentWords.length > 0 && isTeam1 ? 'team-active' : ''}`}
               >
                 <strong>
                   Team 1<h2>{team1Points}</h2>
@@ -88,7 +88,7 @@ const Index = () => {
               </div>
               <div
                 className={`score
-          ${currentWords.length > 0 && !isTeam1 ? "team-active" : ""}`}
+          ${currentWords.length > 0 && !isTeam1 ? 'team-active' : ''}`}
               >
                 <strong>
                   Team 2<h2>{team2Points}</h2>
@@ -101,10 +101,10 @@ const Index = () => {
           <div className="final-score">
             {currentWords.length === 0 &&
               team1Points > team2Points &&
-              "Team 1 Wins 🙌🏻"}
+              'Team 1 Wins 🙌🏻'}
             {currentWords.length === 0 &&
               team2Points > team1Points &&
-              "Team 2 Wins 🙌🏻"}
+              'Team 2 Wins 🙌🏻'}
             {currentWords.length === 0 && (
               <div>
                 <button
@@ -144,9 +144,5 @@ const Index = () => {
     </div>
   );
 };
-
-export async function getServerSideProps(context) {
-  return { props: {} };
-}
 
 export default Index;
